+++ libcnb/examples/example-02-ruby-sample/src/main.rs
@@ -1,50 +1,48 @@
 use std::collections::HashMap;
 use std::process::{Command, Stdio};
 
-use libcnb::data;
+use crate::layers::bundler::BundlerLayerLifecycle;
+use crate::layers::ruby::RubyLayerLifecycle;
 use libcnb::data::build_plan::BuildPlan;
 use libcnb::layer_lifecycle::execute_layer_lifecycle;
-use libcnb::{
-    cnb_runtime, BuildContext, DetectContext, DetectOutcome, GenericErrorHandler, GenericPlatform,
-};
+use libcnb::{cnb_runtime, BuildContext, DetectContext, DetectOutcome, GenericPlatform};
+use libcnb::{data, Buildpack};
 use serde::Deserialize;
-
-use crate::layers::bundler::BundlerLayerLifecycle;
-
-use crate::layers::ruby::RubyLayerLifecycle;
-
 mod layers;
 
+struct RubyBuildpack;
+impl Buildpack for RubyBuildpack {
+    type Platform = GenericPlatform;
+    type Metadata = RubyBuildpackMetadata;
+    type Error = anyhow::Error;
+
+    fn detect(&self, context: DetectContext<Self>) -> libcnb::Result<DetectOutcome, Self::Error> {
+        let outcome = if context.app_dir.join("Gemfile.lock").exists() {
+            DetectOutcome::Pass(BuildPlan::new())
+        } else {
+            DetectOutcome::Fail
+        };
+
+        Ok(outcome)
+    }
+
+    fn build(&self, context: BuildContext<Self>) -> libcnb::Result<(), Self::Error> {
+        println!("---> Ruby Buildpack");
+        println!("---> Download and extracting Ruby");
+
+        let ruby_env = execute_layer_lifecycle("ruby", RubyLayerLifecycle, &context)?;
+
+        println!("---> Installing bundler");
+        install_bundler(&ruby_env)?;
+        execute_layer_lifecycle("bundler", BundlerLayerLifecycle { ruby_env }, &context)?;
+
+        write_launch(&context)?;
+        Ok(())
+    }
+}
+
 fn main() {
-    cnb_runtime(detect, build, GenericErrorHandler)
-}
-
-fn detect(
-    context: DetectContext<GenericPlatform, RubyBuildpackMetadata>,
-) -> libcnb::Result<DetectOutcome, anyhow::Error> {
-    let outcome = if context.app_dir.join("Gemfile.lock").exists() {
-        DetectOutcome::Pass(BuildPlan::new())
-    } else {
-        DetectOutcome::Fail
-    };
-
-    Ok(outcome)
-}
-
-fn build(
-    context: BuildContext<GenericPlatform, RubyBuildpackMetadata>,
-) -> libcnb::Result<(), anyhow::Error> {
-    println!("---> Ruby Buildpack");
-    println!("---> Download and extracting Ruby");
-
-    let ruby_env = execute_layer_lifecycle("ruby", RubyLayerLifecycle, &context)?;
-
-    println!("---> Installing bundler");
-    install_bundler(&ruby_env)?;
-    execute_layer_lifecycle("bundler", BundlerLayerLifecycle { ruby_env }, &context)?;
-
-    write_launch(&context)?;
-    Ok(())
+    cnb_runtime(RubyBuildpack)
 }
 
 #[derive(Deserialize, Debug)]
@@ -68,9 +66,7 @@ fn install_bundler(ruby_env: &HashMap<String, String>) -> anyhow::Result<()> {
     }
 }
 
-fn write_launch(
-    context: &BuildContext<GenericPlatform, RubyBuildpackMetadata>,
-) -> anyhow::Result<()> {
+fn write_launch<B: Buildpack>(context: &BuildContext<B>) -> anyhow::Result<()> {
     let mut launch_toml = data::launch::Launch::new();
     let web =
         data::launch::Process::new("web", "bundle", vec!["exec", "ruby", "app.rb"], false, true)?;